split out `spawn_checksum_build()`
